Move Groupers out to groupers.py
